search backend: clean up search/types.go
